use spring conditions
