+++ instrumentation/spring/spring-boot-autoconfigure/src/main/java/io/opentelemetry/instrumentation/spring/autoconfigure/internal/instrumentation/runtimemetrics/Java17RuntimeMetricsAutoConfiguration.java
@@ -6,13 +6,18 @@
 package io.opentelemetry.instrumentation.spring.autoconfigure.internal.instrumentation.runtimemetrics;
 
 import io.opentelemetry.api.OpenTelemetry;
-import io.opentelemetry.instrumentation.api.incubator.config.internal.InstrumentationConfig;
 import io.opentelemetry.instrumentation.runtimemetrics.java17.RuntimeMetrics;
-import java.util.function.Consumer;
+import io.opentelemetry.instrumentation.spring.autoconfigure.internal.ConditionalOnEnabledInstrumentation;
+import io.opentelemetry.instrumentation.spring.autoconfigure.internal.properties.ConfigPropertiesBridge;
+import io.opentelemetry.sdk.autoconfigure.spi.ConfigProperties;
+import javax.annotation.PreDestroy;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnJava;
 import org.springframework.boot.system.JavaVersion;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
 
 /**
  * Configures runtime metrics collection for Java 17+.
@@ -20,16 +25,31 @@ import org.springframework.boot.system.JavaVersion;
  * <p>This class is internal and is hence not for public use. Its APIs are unstable and can change
  * at any time.
  */
+// we're not using "runtime-telemetry-java17" requirement here, because java17 is just
+// an additional feature - which enables JFR metrics
+@ConditionalOnEnabledInstrumentation(module = "runtime-telemetry")
+@Configuration
 @ConditionalOnJava(value = JavaVersion.SEVENTEEN)
-public class Java17RuntimeMetricsAutoConfiguration implements RuntimeMetricsProvider {
-  private static final Logger logger = LoggerFactory.getLogger(Java17RuntimeMetricsAutoConfiguration.class);
+public class Java17RuntimeMetricsAutoConfiguration {
+  private static final Logger logger =
+      LoggerFactory.getLogger(Java17RuntimeMetricsAutoConfiguration.class);
 
-  @Override
-  public void start(
-      OpenTelemetry openTelemetry, Consumer<Runnable> shutdownHook, InstrumentationConfig config) {
-    logger.debug("Use runtime metrics instrumentation for Java 17+");
-    RuntimeMetrics.builder(openTelemetry)
-        .setShutdownHook(shutdownHook)
-        .startFromInstrumentationConfig(config);
+  private Runnable shutdownHook;
+
+  @PreDestroy
+  public void stopMetrics() {
+    if (shutdownHook != null) {
+      shutdownHook.run();
+    }
+  }
+
+  @Bean
+  CommandLineRunner startMetrics(OpenTelemetry openTelemetry, ConfigProperties configProperties) {
+    return (args) -> {
+      logger.debug("Use runtime metrics instrumentation for Java 17+");
+      RuntimeMetrics.builder(openTelemetry)
+          .setShutdownHook(runnable -> shutdownHook = runnable)
+          .startFromInstrumentationConfig(new ConfigPropertiesBridge(configProperties));
+    };
   }
 }
