+++ instrumentation/spring/spring-boot-autoconfigure/src/main/java/io/opentelemetry/instrumentation/spring/autoconfigure/internal/instrumentation/runtimemetrics/Java8RuntimeMetricsAutoConfiguration.java
@@ -6,15 +6,21 @@
 package io.opentelemetry.instrumentation.spring.autoconfigure.internal.instrumentation.runtimemetrics;
 
 import io.opentelemetry.api.OpenTelemetry;
-import io.opentelemetry.instrumentation.api.incubator.config.internal.InstrumentationConfig;
 import io.opentelemetry.instrumentation.runtimemetrics.java8.RuntimeMetrics;
-import java.util.function.Consumer;
+import io.opentelemetry.instrumentation.spring.autoconfigure.internal.ConditionalOnEnabledInstrumentation;
+import io.opentelemetry.instrumentation.spring.autoconfigure.internal.properties.ConfigPropertiesBridge;
+import io.opentelemetry.sdk.autoconfigure.spi.ConfigProperties;
+import javax.annotation.PreDestroy;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.boot.CommandLineRunner;
 import org.springframework.boot.autoconfigure.condition.AnyNestedCondition;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnJava;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.system.JavaVersion;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Conditional;
+import org.springframework.context.annotation.Configuration;
 
 /**
  * Configures runtime metrics collection for Java 8.
@@ -22,29 +28,42 @@ import org.springframework.boot.system.JavaVersion;
  * <p>This class is internal and is hence not for public use. Its APIs are unstable and can change
  * at any time.
  */
-@ConditionalOnJava(value = JavaVersion.SEVENTEEN, range = ConditionalOnJava.Range.OLDER_THAN)
-public class Java8RuntimeMetricsAutoConfiguration implements RuntimeMetricsProvider {
-  private static final Logger logger = LoggerFactory.getLogger(Java8RuntimeMetricsAutoConfiguration.class);
+@ConditionalOnEnabledInstrumentation(module = "runtime-telemetry")
+@Configuration
+@Conditional(Java8RuntimeMetricsAutoConfiguration.Java8MetricsCondition.class)
+public class Java8RuntimeMetricsAutoConfiguration {
+  private static final Logger logger =
+      LoggerFactory.getLogger(Java8RuntimeMetricsAutoConfiguration.class);
 
-  @Override
-  public void start(
-      OpenTelemetry openTelemetry, Consumer<Runnable> shutdownHook, InstrumentationConfig config) {
-    logger.debug("Use runtime metrics instrumentation for Java 8");
-    RuntimeMetrics.builder(openTelemetry)
-        .setShutdownHook(shutdownHook)
-        .startFromInstrumentationConfig(config);
+  private Runnable shutdownHook;
+
+  @PreDestroy
+  public void stopMetrics() {
+    if (shutdownHook != null) {
+      shutdownHook.run();
+    }
   }
 
-  public class Java8MetricsCondition extends AnyNestedCondition {
+  @Bean
+  CommandLineRunner startMetrics(OpenTelemetry openTelemetry, ConfigProperties configProperties) {
+    return (args) -> {
+      logger.debug("Use runtime metrics instrumentation for Java 8");
+      RuntimeMetrics.builder(openTelemetry)
+          .setShutdownHook(runnable -> shutdownHook = runnable)
+          .startFromInstrumentationConfig(new ConfigPropertiesBridge(configProperties));
+    };
+  }
+
+  static class Java8MetricsCondition extends AnyNestedCondition {
     public Java8MetricsCondition() {
       super(ConfigurationPhase.PARSE_CONFIGURATION);
     }
 
     @ConditionalOnJava(value = JavaVersion.SEVENTEEN, range = ConditionalOnJava.Range.OLDER_THAN)
-    class OlderThanJava17 {}
+    static class OlderThanJava17 {}
 
     // GraalVM native image does not support Java 17 metrics yet
     @ConditionalOnProperty("org.graalvm.nativeimage.imagecode")
-    class GraalVM {}
+    static class GraalVm {}
   }
 }
