--- instrumentation/spring/spring-boot-autoconfigure/src/main/java/io/opentelemetry/instrumentation/spring/autoconfigure/internal/instrumentation/runtimemetrics/RuntimeMetricsProvider.java
@@ -1,22 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.instrumentation.spring.autoconfigure.internal.instrumentation.runtimemetrics;
-
-import io.opentelemetry.api.OpenTelemetry;
-import io.opentelemetry.instrumentation.api.incubator.config.internal.InstrumentationConfig;
-import java.util.function.Consumer;
-
-/**
- * Configures runtime metrics collection.
- *
- * <p>This class is internal and is hence not for public use. Its APIs are unstable and can change
- * at any time.
- */
-public interface RuntimeMetricsProvider {
-
-  void start(
-      OpenTelemetry openTelemetry, Consumer<Runnable> shutdownHook, InstrumentationConfig config);
-}
